Moved testcase to own class
